Use OOP to Access and Change Information/Variables
Previously with global variables, it worked, but we had to
specifically tell each function that the variable was indeed global
using the global keyword to change its value. This isn't the best way
of solving such a problem, especially at scale. It's not very clean.

With classes and OOP, we can solve this same problem more cleanly. It
not only allows us to implement a real world entity as we are doing
here with a bank, but also, instance variables (such as self._balance
in this case) are by definition accessible to all the methods in that
class because we are accessing them all by that special parameter
called self.

Conclusions
- For a reasonably small script where you're trying to implement some
global information such as in this case, the global keyword is a
solution to that problem
- But more generally speaking, in many languages including Python,
using global variables is frowned upon because it tends to become
messy very quickly as you're trying to understand where exactly your
information is stored -- globally or locally, and locally in which
function, etc.
- The rule is to use global variables sparingly (rarely), but it can be
a solution to some of these problems