Introduce Constants
Some languages allow you to define variables that are 'constant'.
I.e. Once you have set a value to them, you cannot change the value of
that variable. This can be a good thing because it allows you to
program defensively, just in case you or someone else accidentally
tries to modify that variable. It cannot be changed, at least not
easily.

This doesn't actually exist in Python. In Python, it is again based on
an honour system -- indicating that someone should be treated as though
it is a constant but it's not actually enforced by the language.

It is best practice to, every time you have a constant -- a number that
just doesn't change, let it bubble up and declare that value as a
variable near the top -- as a constant. This way you and others later
on can more easily recognize this constant. If we ever need to change
the value of the variable in the future, we only need to change one
spot -- where the variable is declared.

In Python, a constant doesn't actually exist, but a variable is to be
treated as a constant if it is in all caps. In other languages, there
is usually a keyword or something that ensures a constant is a constant.