Introduce Use of Class Variables as Constants
Class variables are variables inside a class, that aren't inside of
self itself, but are accessible to all the methods inside of that class.

Here is a convention, not enforced by Python itself, 	of having class
constants. Whereby, inside of a class, you might want to have a variable
 that should not be changed, and you just want to indicate that visually
  by capitalizing that variable's name.

Calling the class constant, you would call the <class>.<class constant>
instead of self.<constant>.
E.g. Cat.MEOWS as seen in this code