Introduce Reason to Use Type Hints
Python is a dynamically typed language (instead of a statically typed
language such as C or Java where you have to explicitly declare the type
of a variable). In Python, you don't have to tell the program what the
type of the variable is. Just give the variable a name and assign it a
value and Python will figure out what type of variable it will be.

The advantage of a statically typed language is that it helps to detect
bugs more readily. I.e. you know what type a variable should and
shouldn't be. Whereas with dynamically typed languages, you wouldn't
have as much control and so wouldn't as easily know what type a variable
is. In fact, for statistically typed languages, the language itself
can detect that a wrong type is being used.

In Python, you can provide hints as to what type a variable should be,
but the language itself does not strongly enforce these. You can use a
tool to tell you whether you're using a variable correctly or not, but
it's typically a tool you would run as a programmer before you release
your code to the world.

There is a popular program for checking whether or not your code is
adhering to your own type hints. Specifically, the program is mypy.

In this example, this leads to a TypeError because input returns a str,
and we are trying to pass an int as the argument to the meow() function.