Using mypy with Type Hints
Include a type hint in the parameter of a function or for any variable.
Format -- <variable>: <type>
E.g. number: int = ...

This doesn't actually do anything as it isn't enforced by the Python
language itself, but it gives a small indication, or a hint, on what
type the arguments passed into the function should be. Additionally,
it allows other programs, such as mypy, to better work with your script
and the type hint and more prominently indicate to you the errors.

Run using "mypy <Python file name>.py" in terminal.