Introduce Type Hint for Function Return Values
By default, when a function in Python doesn't return a value, its
implicit return value is None.

In this example, the function doesn't return a string but instead
returns None and has the side effect of printing. The None return
value is indicated by the return value type hint.
The format for this is:
<function> -> <type hint>:

However, outside of the function, we're expecting the function to
return a string, as indicated by the type hint for the variable. With
the type hint for the return value and the type hint for the variable
meow being different, mypy will catch this error and indicate that to
the programmer.