Introduce Command Line Argument Flags with If/Else
It is common convention to print out "usage: <how to use program>" when
user is incorrectly using the program.

When controlling programs from the command line, it is common to provide
switches or flags whereby you pass in something such as "-n" (which
means this <number> of times) then space and then the number.
E.g. python <program> -n 3

This still allows other things to be done at the command line, but now
it's standardized on how command line arguments are provided to the
program. It's more reliable for the program to know what the 3 means.
I.e. vs. python <program> 3
- In this case, the 3 is just there. It's a bit ambiguous and less
obvious, and we know what it means in this case, but it could really
mean anything if we didn't know what it already meant, as compared to
flagging it with -n which would slightly more obviously indicate that
3 is a number to be passed into the program.

However, imagine if this program gets more complicated where we start
supporting not only -n but also -a, -b, -c, etc. At that point, it
might even be helpful to use words, such as --number. It is a convention
in computing to use single dashes with a single letter (-n) but use
double dashes if you're using a whole number (--number). This gets
really complicated to do with if/else statements. You're going to have
to check all over the place. And what if users input the flags in
different orders?
This problem has already been solved with the library argparse.