Introduce argparse
Argparse is a library that handles all of the parsing and analysis of
command line arguments for you. This allows the program to abstract
away this problem of command line arguments and focus more on the
problem that the programmer is actually trying to solve.

Don't need to import sys because argparse would have imported sys
already and will handle everything.

parser.add_argument("-n")
- This adds -n as an acceptable command line argument for the parser
to recognize

args = parser.parse_args()
- Here, the parser parses all of the command line arguments
- And then the result creates and stores in an object called args,
inside of which would be all of the values of the command line arguments
 no matter which order they appeared in

args.n
- This accesses the n flag from the command line argument within the
args object

However, this results in a TypeError if we just run "python meow.py"
instead of "python meow.py -n <number>". This is because we haven't
accounted for this in our code yet.