Add Usage Info. & Error Handling via Argparse
It is conventional in Python (and other languages) to run a program
with a special argument: -h or --help. What should do is provide the
user with usage information -- how to use the program.

In documentation square brackets [ ] means the parameter(s) is optional.

When we type "python meow.py -h", we'll see:
"-n N        number of times to meow"
- The N after the -n here indicates that the program expects a number
after the -n flag

[Q&A]
What happens if the user does not type an int? How does the program,
specifically argparse, handle this error?
- It prints out an error message with a brief explanation reminding
the user how to use the program