Try to Pass One Variable for 3-Parameter Function
This mistake might come up for those trying to pass in a list of values
into a function to try to satisfy the number of arguments the function
accepts. Specifically in this case, there is a list of 3 values for a
function that accepts 3 arguments. However, what we're actually doing
here is passing a list of values to the galleons argument and nothing
for sickles or knuts. In other words, we're passing in only 1 argument
when the function expects 3. This is especially highlighted by how the
function doesn't have default values for the parameters. If it had
default values, then this could have been accepted, but the code would
have to be changed so that it handled this specific case.

It's also the wrong type. We're passing in a list when the galleons
parameter expects an integer as that's what the galleons variable will
be used as in the total() function.

This leads to a TypeError and highlights that the function total() is
missing 2 required positional arguments.

However, what we shouldn't be doing is
total(coins[0], coins[1], coins[2]) as this would be very verbose and
vulnerable to human mistakes in the form of typos.

Here too, we can actually do some unpacking.