Unpack a List
A list is packed with values. In this case, it's packed with 100, 50,
25. In Python, this can be unpacked by adding an asterisk/star (*) at
the beginning of the variable. In this case, we are unpacking the list
into 3 individual values which can then be appropriately passed into
our function which is expecting 3 individual arguments.

Overall, the effect of passing *coins is to pass in the individual
values of that list (100, 50, 25) into the total() function. Our code
is also much cleaner compared to doing coin[0], coin[1], coin[2].

[Q&A]
Does this work for tuples, sets, dictionaries?
- Tuples -- yes
- Sets -- no
- Dictionaries -- there's a different way
- It will work for enumerations in which order is preserved

Can we do 100+50+25 directly in this case?
- No
- Unpacking in this way returns the individual values which is similar
to splitting from before and setting them equal to individual variables.
 You can't directly manipulate them, such as adding them, in this way.