Change to Passing in Named Arguments
In this case, this is no longer positional -- i.e. the position of the
arguments don't matter. This is because we're now manually passing in
the values and explicitly specifying the names of these arguments. By
specifying the names of the arguments according to the parameters of
the function, we are able to set what each specific argument is.