Use a Dictionary
Dictionaries are collections of key-value pairs; i.e. names and values.
Therefore, instead of just passing in names values as they are, we can
instead create a dictionary with key-value pairs and pass in those
individual pairs as the arguments for the total() function.

However, this runs into a similar problem to coins[0], coins[1],
coins[2]. It is verbose. There is a similar solution to unpacking
lists but with dictionaries.