Unpack a Dictionary Using **
For a list, we used one asterisk (*) -- e.g. *coins.

For a dictionary, it is similar. We use two asterisks (**)
-- e.g. **coins.
What this will do is pass in 3 values with names. It has the effect as
passing in named arguments as before.
I.e. **coins will pass in "galleons=100, sickles=50, knuts=25" into the
total() function
Therefore, the keys have to match the names of the parameters of the
function. If not, there will be a TypeError indicating that there is an
unexpected keyword argument.

[Q&A]
For both unpacking a list and dictionary and passing those into a
function, the number of items unpacked have to match the number of
items expected as arguments. If not, there will be a TypeError. At
least for how we've implemented our functions here. It is possible to
define functions that take in variable number of arguments as will be
seen later.