Accept a Variable Number of Positional Arguments
Single (*) and double (**) asterisks aren't only used in unpacking, but
also as a visual indicator in Python for when a function takes a
variable number of arguments. I.e. a function can be variadic. This
means, it doesn't necessarily have to take 3 arguments specifically,
even if they do or don't have default values. They can take
0, 1, 2, or 4 arguments as well. The syntax for implementing this idea
is shared with unpacking.

Syntax:
*args
- Indicates that the function takes a variable number of POSITIONAL
arguments
- It's called args here to represent arguments and is the convention,
but can technically be named anything

**kwargs
- Indicates that the function takes a variable number of NAMED
arguments; i.e. named parameters that can be called optionally
individually by their own names
- It's called kwargs here to represent keyword arguments and is the
convention, but can technically again be named anything

*args is automatically a tuple.
When a tuple has only 1 element, this is indicated by a tuple () with
only one element followed by a comma, indicating that it is indeed a
tuple