Demonstrate print() Uses Variable Number of *args
So, we've actually variable number of arguments before in print(). The
documentation for the print() function shows:
print(*objects, sep=' ', end='\n', file=sys.stdout, flush=False)
- What we can infer from what we've learned from the use of asterisk
today is that print() takes a variable number of positional arguments
via the *objects parameter
- Therefore, in the print() function, there is probably a loop that
iterates over each of the (variable number of) objects, allowing the
print() function to print each of those one at a time

Based on what we know, this is the possible code for how print() was
implemented by the authors of Python

[Q&A]
Can we pass *args or **kwargs from one function into another function?
- Yes
- You might want to do this if you want to "wrap" another function --
provide some additional functionality but still pass in all of the
supported arguments to the underlying function as well