Create Basic yell.py & Programming Paradigms
[Review of programming paradigms (with respect to Python)]
Procedural programming
- Did this in the beginning
- Wrote code top to bottom, left to right, defining some functions
(procedures) along the way, defining variables, and having side effects
and assigning values as needed

Object-oriented programming
- Python is very much object-oriented
- A lot of the variables, types were actually objects
- Objects come from certain classes
- Classes were templates/blueprints by which you can encapsulate both
data and functionality

Functional programming
- Saw hints of this along the way
- With functional programming, functions are more powerful where they
tend NOT to have side effects (no printing or changing of states
globally), but rather they are completely self contained and might take
functions themselves as inputs and return values
- We saw this paradigm when we sorted things, particularly with the
sort() function and lambda function. We passed in the function we wanted
 to use to sort a list/dictionary
- Python also has other functionalities that is reminiscent of
functional programming that is a powerful way to solve problems a
little more differently still. This is called mapping.