Introduce map() and More Functional Programming
map() is a function in Python (and other languages) that allows you to
take a list or a sequence of values and "map" a function to each of
those values and get back all of those return values. map() allows you
to apply some function to every element of some sequence, like a list.

"map" is a function, but it is also a verb. You can map a function to a
sequence of values. That is, line the function up with each of the
values so that you're calling the function on each value.

In this case, if we want to force to uppercase each phrase, we would
want to map the uppercase function to each of those values (phrases).

map() accepts two arguments:
- the name of a function to map. When including this function as an
argument, we don't actually call the function. We don't want to call
the function now. The map() function will call the function for us when
it needs to. Therefore, when including the function as an argument, we
omit the parentheses () and just include the function's name.
- the sequence of values to map the function to (an iterable).
E.g. map(str.upper, words)

The use of map() here passes one function to another. This is a hallmark
of functional programming where we can treat functions as objects
themselves; values that can be passed around functions.