Introduce List Comprehensions
Pythonic/common in Python.

List comprehension refers to the ability in Python to easily construct
a list on the fly -- without using a loop, without calling append().
Everything is done in one line.

Specifically in Python, list comprehension is the opportunity to create
a list using square brackets, but inside those square brackets, instead
of values, is a Python expression that in effect is going to
dynamically generate a brand new list based on the logic written.

E.g. uppercased = [word.upper() for word in words]
- There is a value/function call on the left and loop on the right
- How to read this:
"give me the uppercase version of the word for each word in words list"

[Q&A]
Can we include conditionals in the list comprehension, such as if-else?
- Yes.

Is this functional programming?
- Not really, just a feature of Python

Aspects of functional programming so far
- Lambda functions and passing them in as a key attribute to the
sorted() function
- map() function where we're passing in a function to map values to
- We'll see another one in a bit