Introduce filter()
A functional approach to filtering a list result.

filter(<function>, <iterable>)
- filter(), similar to map(), takes two arguments:
1) name of a function that will return a boolean value (True/False);
again, don't call the function, so don't include the parentheses ()
2) an iterable/sequence with elements to which the function will be
applied
- Difference is, map() returns one value for each element for the
sequence, but filter() applies a conditional so it only returns the
value/element from the sequence if it matches the condition

This code also includes a lambda function for the sorted() function to
sort the dictionary

So overall, there were two approaches to this problem of "filtering"
out a list of results from another list.
1) list comprehension with if conditional
2) filter() function -- more functional approach

[Q&A]
In filter(), instead of writing a separate function that you would only
use once just for the filter() function, can you write a function
directly in the filter() function?
- Yes. You can write a lambda function instead