Introduce Reason to Use Dictionary Comprehensions
With dictionary comprehensions, we have the ability to create a
dictionary on the fly, with keys and values, without having to doing it
the longer normal way -- creating an empty dictionary, creating a for
loop and iterating over that loop to insert specific keys. Instead with
dictionary comprehensions, we can just do it all at once.

In this code, we have built a list of dictionaries the longer "normal"
way.