Introduce List Comprehension for List of Dicts
Inside of this list comprehension, we're indicating that we want a
dictionary (as indicated by the curly braces {}) that has two keys --
"name" and "house" respectively with the values of the students names
and "Gryffindor. Which students do we want to create those dictionaries
from? -- for student in students.
- Therefore, on the left is what we want in the final list and on the
right is the loop. No conditional this time since we want all students
in our final list.

What if we don't want a list of dictionaries but just one big
dictionary?