Introduce Dictionary Comprehension
Here, we just want one big dictionary where they keys are the students'
names and the values are their house. We are assuming that no one will
have the same first name in our dictionary.

The curly braces {} here indicate that we want a dictionary (vs. Square
brackets [] for list before), and we indicate what every key and value
will look like based on a for loop, similar to list comprehension
before.