Introduce Reason to Use enumerate()
If we want to print out a student's "ranking" or a numbered list, we
would normally have to do a "for i in range(len(<list>))" loop
(accounting for the fact that i starts at 0 with a +1).

However, Python actually has a slightly easier way to do this with the
enumerate() function.