Refactor sheep String into a Helper Function
Reason to use helper functions (vs. Keeping everything in main()):
- Abstraction - no longer have to think about how they're implemented
- Easier to test in unit tests

We now at least have a framework to focus on the implementation of this
sheep function.

However now, it's not very elegant to still have main() function to do
the iteration. Wouldn't it be nice to define a function to return the
entire string that we want rather than just one row of sheep at a time?