Refactor Code to Return List of Sheep
Return the entire flock/list of sheep instead of returning one row at a
time. This way, main() only has to handle the printing after, and
sheep() will handle the production of the sheep including the individual
rows and looping through and creating every row of sheep.

We run into a problem this way when we run a function that will create
and return the entire flock of sheep if that flock is a very large
number. We begin pushing our computer's capabilities. After some
threshold, it will overload our computer and cause the computer to kill
the program. The essence of this problem is again that the function
waits until the entire flock is created and then returns that entire
flock. However, we shouldn't do it as before either where our function
would only return one line at a time and the loop is in the main()
function. We generally don't want to put everything in main() and it
would be good to have separate functions for abstraction and for unit
testing purposes. These are our best practices, but then we run into
this problem. What's the solution?