Introduce Python Generator Functions and yield
In Python, we can define a function as a <generator> whereby it can
still generate a massive amount of data for your users, but you can
have it just return a little bit of that data at a time.
Fundamentally, it all boils down to the keyword <yield>.

yield
- Up until now when defining functions, we have been defining functions
that return values, if at all, using the keyword return. In our
previous example, that was what we did. We waited until our function
ran through everything to return everything all at once. This
ultimately led to our program to be killed by the computer because the
program used too much memory and CPU.

In our sheep() function, if we had done what we would normally
previously do in the middle of our loop:
return "🐑" * i
- This would have only returned to the main() function the very first
line of the loop and ended the loop there. We wouldn't get the
subsequent iterations of the loop.
- But, thanks to yield, we can tell Python to just return one
value/iteration at a time from the loop in our function.
- Therefore, each time the function will just hand back/generate a
little piece of data rather than everything all at once.

Yield is technically returning an <iterator> that allows our own
code/for loop in main() to iterate over these generated values one at
a time.

[Q&A]
How does yield actually work under the hood?
- You can think of it as being asynchronous in a sense, whereby
returning a value immediately and then subsequently giving you another
and more for each iteration
- Underneath the hood, what's really happening is the generator is
retaining state for you. It does not run the entire loop from top to
bottom and then return a value one a time, but instead, it does one
iteration and yield the result, and Python will suspend the function
but also remember what iteration it was on so that the next time you
iterate over it, as you would in the for loop in main(), it will just
continue and you'll get back the next value from the loop in the
function again and again
- Yield gets back an iterator and the iterator can be stepped over as
in a loop one element at a time, but Python handles all of that for you
so you don't need to do any of the underlying coding

Since the program returns over one iteration at a time vs before where
everything was returned all at once, if we wanted to print a million
rows of sheep, this is in a sense 1 millionth the amount of
memory (oversimplified)